% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dDirichlet.R
\name{dDirichlet}
\alias{dDirichlet}
\title{A dDirichlet Function}
\usage{
dDirichlet(x, alpha, log = FALSE)
}
\arguments{
\item{x, }{Input parameter}

\item{alpha, }{Hyper-prior}

\item{log=False, }{use log?}
}
\description{
density of dirichlet
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))
}
\keyword{dirichlet}

