% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dMvn.R
\name{dMvn}
\alias{dMvn}
\title{Multivariate normal density, a bit faster}
\usage{
dMvn(X, mu, Sigma)
}
\arguments{
\item{x}{mu Sigma}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
# dMvn
}
\keyword{multivariate}

