% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layOut.R
\name{layOut}
\alias{layOut}
\title{PostProcessing Predictive function}
\usage{
layOut(...)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

