% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot function and ggplot2}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist=NULL, }{file, cols=1}
}
\description{
Multiplot function and ggplot2
}
\examples{
# used inside other functions
}
\keyword{ggplot}

