% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostPredFunk.R
\name{PostPredFunk}
\alias{PostPredFunk}
\title{PostProcessing: Predictive function. Used in main Post-processing function}
\usage{
PostPredFunk(.GrunK0us, .Zetc, .Y, .prep, .simlabel)
}
\arguments{
\item{.GrunK0us, }{Unswitched MCMC output for one _K0}
}
\description{
This function generates new datasets from MCMC output of ZMIX.
}
\examples{
#nope
}
\keyword{PostProcessing}
\keyword{Predictive,}

