% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PostProcUniv_Clusters.R
\name{PostProcUniv_Clusters}
\alias{PostProcUniv_Clusters}
\title{PostProcessing function univ}
\usage{
PostProcUniv_Clusters(Grun, mydata, LineUp = 1, Propmin = 0.05,
  isSim = TRUE, nEnd = 2000)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

