<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> parent of cc5b841... Major Updates
% Please edit documentation in R/PostProcUnivariate_later.R
\name{PostProcUnivariate_later}
\alias{PostProcUnivariate_later}
\title{PostProcessing function univ}
\usage{
PostProcUnivariate_later(Grun, mydata, LineUp = 1, prep = 10000,
  Propmin = 0.05, isSim = TRUE, simlabel = "sim", savelabel = "PPplot",
  nEnd = 2000)
}
\arguments{
<<<<<<< HEAD
\item{Grun, }{mydata, LineUp, prep, Propmin, isSim, simlabel, savelabel, nEnd.}
=======
\item{Grun,}{mydata, LineUp, prep, Propmin, isSim, simlabel, savelabel, nEnd.}
>>>>>>> parent of cc5b841... Major Updates
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

