% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PostProcAwesome.R
\name{ProstProcAWESOME}
\alias{ProstProcAWESOME}
\title{PostProcessing function univ}
\usage{
ProstProcAWESOME(Grun, Y, prep = 10000, LineUp = 1, Propmin = 0.3,
  simlabel = "sim")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

