% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zswitch_multivariate.R
\name{QuickSwitchMVN}
\alias{QuickSwitchMVN}
\title{Label switch fix for Multivariate Normals}
\usage{
QuickSwitchMVN(MVNResult, PropMin = 0.3)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#not run
}
\keyword{Wishart}

