% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{A Dirichlet Function}
\usage{
rdirichlet(m = 1, par)
}
\arguments{
\item{m}{Vector of Alpha values (hyperparameter of dist)}
}
\description{
This function draws samples from q dirichlet distriution with hyperpars m.
}
\examples{
rdirichlet()
}
\keyword{dirichlet}

