% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Replicated_Zmix_light.R
\name{ReplicatedZmix_light}
\alias{ReplicatedZmix_light}
\title{A replicated  Function}
\usage{
ReplicatedZmix_light(NumRep, sim, n, K = 10, mylabels = "trial",
  nIter = 20000)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
...
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))


 go<- ReplicatedZmix_light(3, 1, 20, K=5, iter=400)
}
\keyword{dirichlet}

