% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReplicatedZmix_parallelQUICK.R
\name{ReplicatedZmix_parallelQUICK}
\alias{ReplicatedZmix_parallelQUICK}
\title{A replicated  Function}
\usage{
ReplicatedZmix_parallelQUICK(NumRep, sim, n, mylabels = "trial", ...)
}
\arguments{
\item{x,}{alpha, log=False}
}
\description{
...
}
\examples{
dDirichlet(c(.1, .9), c(0.1,0.1))


 go<- ReplicatedZmix_light(3, 1, 20, K=5, iter=400)
}
\keyword{dirichlet}

