% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rinvgamma.R
\name{rinvgamma}
\alias{rinvgamma}
\title{Draw from Inverse Gamma}
\usage{
rinvgamma(n, shape, scale = 1)
}
\arguments{
\item{m}{Vector of Alpha values (hyperparameter of dist)}
}
\description{
This function draws samples from an inverse Gamma dist
}
\examples{
rinvgamma(1, 2)
}
\keyword{Gamma}

