% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rwish.R
\name{rwish}
\alias{rwish}
\title{Draw from Wishart}
\usage{
rwish(v, S)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
rwish(2,1)
}
\keyword{Wishart}

