% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim1func}
\alias{sim1func}
\title{Sim 1 function}
\usage{
sim1func(n)
}
\arguments{
\item{n:}{Size}
}
\description{
Simulation function for Zmix paper.
}
\examples{
#NA
}
\keyword{simulation}

