% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim2EASYfunc}
\alias{sim2EASYfunc}
\title{Sim 2 k EASY}
\usage{
sim2EASYfunc(n)
}
<<<<<<< HEAD
=======
\arguments{
\item{n}{}
}
>>>>>>> parent of cc5b841... Major Updates
\description{
This function draws samples from a simulation
}
\examples{
<<<<<<< HEAD
#nope			
=======
#nope
>>>>>>> parent of cc5b841... Major Updates
}
\keyword{simulation}

