% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim2func}
\alias{sim2func}
\title{Sim 2 function}
\usage{
sim2func(n)
}
\arguments{
\item{n:}{size}
}
\description{
Simulation function for Zmix paper.
}
\examples{
#NA			
}
\keyword{simulation}

