% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim3func}
\alias{sim3func}
\title{Sim 3 function}
\usage{
sim3func(n)
}
\arguments{
\item{n:}{size}
}
\description{
Simulation function for Zmix paper.
}
\examples{
#NA
}
\keyword{simulation}

