% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim6func}
\alias{sim6func}
\title{Sim 6 function}
\usage{
sim6func(n)
}
\arguments{
\item{n:}{size}
}
\description{
This function draws samples from a simulation
}
\examples{
#NA
}
\keyword{simulation}

