% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{simMe}
\alias{simMe}
\title{Sim Big func}
\usage{
simMe(whichOne = 1, n)
}
<<<<<<< HEAD
=======
\arguments{
\item{n}{}
}
>>>>>>> parent of cc5b841... Major Updates
\description{
This function ...
}
\examples{
#nope
}
\keyword{simulation}

