% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimMVN.R
\name{SimMVN}
\alias{SimMVN}
\title{Simulate multivariate normal Mixture}
\usage{
SimMVN(Means, Covs, N, P)
}
\arguments{
\item{Means,}{covariance, Sample size, weights}
}
\description{
...
}
\examples{
#not run
}
\keyword{Wishart}

