% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simudZ.R
\name{simudZ}
\alias{simudZ}
\title{Simulate Univariate normal Mixture}
\usage{
simudZ(n, mu, sig, p, k)
}
\arguments{
<<<<<<< HEAD
\item{n, mu, sig, p, k}{}
=======
\item{n,mu,sig,p,k}{}
>>>>>>> parent of cc5b841... Major Updates
}
\description{
...
}
\examples{
#not run
}
\keyword{Wishart}

