% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimit.R
\name{trimit}
\alias{trimit}
\title{trim univariate gibbs output}
\usage{
trimit(Out = Out, nEnd = EndSize)
}
\arguments{
\item{Out:}{Output of Zmix}

\item{nEnd:}{How many iterations to KEEP? (i.e. end size)}
}
\description{
This function is part of post processing function
}
\examples{
#NA
}
\keyword{mcmc}
\keyword{trim,}

