% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zagg.R
\name{Zagg}
\alias{Zagg}
\title{Processing output of label switching function}
\usage{
Zagg(USout, .Y = Y)
}
\arguments{
\item{USout:}{output of Zswitch}

\item{.Y:}{Y, vector of original data}
}
\description{
explain here
}
\examples{
#nope
}
\keyword{postprocessing}

