% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zmix_light.R
\name{Zmix_light}
\alias{Zmix_light}
\title{Run Gibbs sampler with prior tempering}
\usage{
Zmix_light(y, k, iter = 5000, isSim = TRUE, burn = 300, alphas = c(30,
  20, 10, 5, 3, 1, 0.5, 1/2^(c(2, 3, 4, 5, 6, 8, 10, 15, 20, 30))))
}
\arguments{
\item{y,}{k,iter, isSim=TRUE, alphas=}
}
\description{
This function ...
}
\examples{
#... you know...
}
\keyword{Gibbs}
\keyword{estimation}
\keyword{gaussian,}
\keyword{mixture,}
\keyword{normal,}
\keyword{order}
\keyword{parallel}
\keyword{sampler,}
\keyword{tempering,}
\keyword{univariate,}

