% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zmix_lightLYRA.R
\name{Zmix_lightLYRA}
\alias{Zmix_lightLYRA}
\title{Run Gibbs sampler with prior tempering}
\usage{
Zmix_lightLYRA(y, iter = 20000)
}
\arguments{
\item{y,}{k,iter, isSim=TRUE, alphas=}
}
\description{
This function ...
}
\examples{
#... you know...
}
\keyword{Gibbs}
\keyword{estimation}
\keyword{gaussian,}
\keyword{mixture,}
\keyword{normal,}
\keyword{order}
\keyword{parallel}
\keyword{sampler,}
\keyword{tempering,}
\keyword{univariate,}

