% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zmix_multi_tempered_withTracker.R
\name{Zmix_multi_tempered_OLD}
\alias{Zmix_multi_tempered_OLD}
\title{This functionZmix_multi_tempered}
\usage{
Zmix_multi_tempered_OLD(YZ, iter, k, alphas, sim = TRUE, EndSize = 500)
}
\arguments{
\item{stuff}{and more}
}
\description{
This functionZmix_multi_tempered
}
\examples{
#not run
}
\keyword{multi}

