% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Zmix_switchByK0.r
\name{Zmix_switchByK0}
\alias{Zmix_switchByK0}
\title{PostProcessing function univ}
\usage{
Zmix_switchByK0(Grun, mydata, prep = 10000, LineUp = 1, Propmin = 0.05,
  isSim = TRUE, simlabel = "sim")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

