% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layOut.R
\name{layOut}
\alias{layOut}
\title{Utility for plotting multiple plots in one frame}
\usage{
layOut(...)
}
\arguments{
\item{inputs, }{NA}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#used inside functions
}
\keyword{plots}

