% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Create multiplot ggplot2}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plots}{and rest}
}
\description{
This function ....
}
\examples{
# not run
}
\keyword{ggplot}

