% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PostPredFunk.R
\name{PostPredFunk}
\alias{PostPredFunk}
\title{PostProcessing Predictive function}
\usage{
PostPredFunk(.GrunK0us, .Zetc, .Y, .prep, .simlabel)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

