% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PostProcUnivariate_later.R
\name{PostProcUnivariate_later}
\alias{PostProcUnivariate_later}
\title{PostProcessing function univ}
\usage{
PostProcUnivariate_later(Grun, mydata, LineUp = 1, prep = 10000,
  Propmin = 0.05, isSim = TRUE, simlabel = "sim", savelabel = "PPplot",
  nEnd = 2000)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

