% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Process_Output_Zmix.R
\name{Process_Output_Zmix}
\alias{Process_Output_Zmix}
\title{Process the output of Zmix_univ_tempered}
\usage{
Process_Output_Zmix(Grun, isSim = FALSE, Burn = 2000, LineUp = 1,
  Pred_Reps = 100, Zswitch_Sensitivity = 0.01, makePlots = TRUE,
  Plot_Title = "Results", SaveFileName = "zmix", PlotType = "Boxplot")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

