% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim4func}
\alias{sim4func}
\title{Sim 4 function}
\usage{
sim4func(n)
}
\arguments{
\item{n:}{size}
}
\description{
Simulation function for Zmix paper.
}
\examples{
#NA
}
\keyword{simulation}

