% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{sim6func}
\alias{sim6func}
\title{Sim 6 function}
\usage{
sim6func(n)
}
\arguments{
\item{n}{}
}
\description{
This function draws samples from a simulation
}
\examples{
#nope
}
\keyword{simulation}

