% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{simMe}
\alias{simMe}
\title{Sim Big function}
\usage{
simMe(whichOne = 1, n)
}
\arguments{
\item{n:}{size}
}
\description{
used in Zmix paper for generating replicate simulations
}
\examples{
#NA
}
\keyword{simulation}

