% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudZ.R
\name{simudZ}
\alias{simudZ}
\title{Simulate data from a Univariate normal Mixture}
\usage{
simudZ(n, mu, sig, p, k)
}
\arguments{
\item{n:}{sample size to simulate}

\item{mu:}{vector of mixture means}

\item{sig:}{vector of mixture component variances}

\item{p:}{vector of mixture weights}

\item{k:}{Number of components.}
}
\description{
...
}
\examples{
simudZ(n=10, mu=c(-1,4,10), sig=c(1,1,2), p=c(.2, .7, .1), k=3)
}
\keyword{Gaussian,}
\keyword{Simulate,}
\keyword{Univariate}

