% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simudZ.R
\name{simudZ}
\alias{simudZ}
\title{Simulate Univariate normal Mixture}
\usage{
simudZ(n, mu, sig, p, k)
}
\arguments{
\item{Means,}{covariance, Sample size, weights}
}
\description{
...
}
\examples{
#not run
}
\keyword{Wishart}

