% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimit.R
\name{trimit}
\alias{trimit}
\title{trim univariate gibbs output}
\usage{
trimit(Out = Out, nEnd = EndSize)
}
<<<<<<< HEAD
=======
\arguments{
\item{stuff}{}
}
>>>>>>> parent of cc5b841... Major Updates
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{stuff}

