% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Zagg.R
\name{Zagg}
\alias{Zagg}
\title{Processing output of label switching function}
\usage{
Zagg(USout, .Y = Y)
}
\arguments{
\item{output}{of UNIVARIATE label switching function}
}
\description{
explain here
}
\examples{
#nope
}
\keyword{postprocessing}

